requirements.txt
motor
roblox
discord.py
pyyaml